requirements.txt
altair==5.5.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.2
cycler==0.12.1
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1744321358928/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
fonttools==4.57.0
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1743437245292/work
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1733462582183/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1744033187/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219691421/work
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
narwhals==1.37.1
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numba==0.61.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1742254834849/work/dist/numpy-2.2.4-cp312-cp312-macosx_10_13_x86_64.whl#sha256=2372f39a2a82db88c99a9180a56ea214161625b167a8433118c3132e031d39f1
optree @ file:///Users/runner/miniforge3/conda-bld/optree_1744034229176/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1744430514379/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
prince==0.16.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1740663229826/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==19.0.1
pybind11 @ file:///D:/bld/pybind11-split_1730237329882/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1730237328107/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1743831293069/work
referencing==0.36.2
rpds-py==0.24.0
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1736496783963/work/dist/scikit_learn-1.6.1-cp312-cp312-macosx_10_13_x86_64.whl#sha256=0a7b8f8bfd367eb107dbc11a09420ca1a4d8e263d0b5e380a45b190f689a8364
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1739790672351/work/dist/scipy-1.15.2-cp312-cp312-macosx_10_13_x86_64.whl#sha256=49a892575264629624d68a6ac2cdbbf2547aedab92d64742c7c0fda1418ccafb
seaborn==0.13.2
setuptools==75.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1727529700862/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
torch @ file:///Users/runner/miniforge3/conda-bld/libtorch_1744233438518/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1732615945596/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
umap-learn==0.5.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
wheel==0.45.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
